move reusable ldap code to helper
