+++ builtin/credential/ldap/backend_test.go
@@ -654,23 +654,6 @@ func testAccStepLoginNoGroupDN(t *testing.T, user string, pass string) logicalte
 	}
 }
 
-func TestLDAPEscape(t *testing.T) {
-	testcases := map[string]string{
-		"#test":       "\\#test",
-		"test,hello":  "test\\,hello",
-		"test,hel+lo": "test\\,hel\\+lo",
-		"test\\hello": "test\\\\hello",
-		"  test  ":    "\\  test \\ ",
-	}
-
-	for test, answer := range testcases {
-		res := EscapeLDAPValue(test)
-		if res != answer {
-			t.Errorf("Failed to escape %s: %s != %s\n", test, res, answer)
-		}
-	}
-}
-
 func testAccStepGroupList(t *testing.T, groups []string) logicaltest.TestStep {
 	return logicaltest.TestStep{
 		Operation: logical.ListOperation,
