+++ helper/ldaputil/config_test.go
@@ -1,0 +1,20 @@
+package ldaputil
+
+import "testing"
+
+func TestLDAPEscape(t *testing.T) {
+	testcases := map[string]string{
+		"#test":       "\\#test",
+		"test,hello":  "test\\,hello",
+		"test,hel+lo": "test\\,hel\\+lo",
+		"test\\hello": "test\\\\hello",
+		"  test  ":    "\\  test \\ ",
+	}
+
+	for test, answer := range testcases {
+		res := escapeLDAPValue(test)
+		if res != answer {
+			t.Errorf("Failed to escape %s: %s != %s\n", test, res, answer)
+		}
+	}
+}
